Made a semi-functional preview route (view.html).
app.py: added default positions needed to be obtained automatically. Also added a test route that can be deleted since I am using port 8000 now and it is stable

gui.py: - Added libraries and function to export image as base64 to be rendered in the webapp.
- Replaced the widgets of the show() method with dictionaries to ensure basic functioning of the view route. Will need to delete them in the future.
- Forced the get_channel_image method to get the first FOV. Need to change this to make it responsive to a slider.

view.html: AI created a base code that needs to be adapted with the corresponding widgets.